Added device info and brightness.

+++ firmware/app_layer_v1/button.h
@@ -13,11 +13,12 @@
 
 #else // ENABLE_BUTTON
 
-#include <stdint.h>
+//#include <stdint.h>
 
 // Initialize this module.
 void ButtonInit();
-uint8_t ButtonPressed();
+
+int ButtonPressed();
 
 
 #endif // ENABLE_BUTTON
